+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Highlights are marked with a pancake 🥞
 
 - Remove pruning logic in LWW TODO example [#1172](https://github.com/p2panda/p2panda/pull/1172)
 - Fix premature termination of p2panda-sync stream on duplicate events [#1182](https://github.com/p2panda/p2panda/pull/1182)
+- Export missing `DecodeError` and `ReplayError` [#1183](https://github.com/p2panda/p2panda/pull/1183)
 
 ## [0.6.0] - 18/05/2026